Move EncounterRenderer to own class
